# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SubscribeInput {
  email: string;
  firstName: string;
  businessName: string;
  phone: string;
  category: string;
  agreedToTerms: boolean;
}

export interface ContactInput {
  name: string;
  email: string;
  subject?: string;
  message: string;
  /** Honeypot field — must be left empty by real users */
  website?: string;
}

export interface ContactResult {
  success: boolean;
  message: string;
}

export interface SubscribeResult {
  success: boolean;
  message: string;
}

export interface ErrorResult {
  error: string;
}