# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: subscribe
    description: Email subscription operations
  - name: contact
    description: Contact form operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /subscribe:
    post:
      operationId: subscribe
      tags: [subscribe]
      summary: Subscribe a lead to Sender.net
      description: Captures lead info and forwards it to Sender.net via their API v2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscribeInput"
      responses:
        "200":
          description: Successfully subscribed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscribeResult"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResult"
        "500":
          description: Server or upstream error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResult"
  /contact:
    post:
      operationId: contact
      tags: [contact]
      summary: Send a contact form message
      description: Forwards a contact form message to the business inbox via Gmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "200":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactResult"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResult"
        "500":
          description: Server or upstream error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResult"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    SubscribeInput:
      type: object
      required:
        - email
        - firstName
        - businessName
        - phone
        - category
        - agreedToTerms
      properties:
        email:
          type: string
        firstName:
          type: string
        businessName:
          type: string
        phone:
          type: string
        category:
          type: string
        agreedToTerms:
          type: boolean
    ContactInput:
      type: object
      required:
        - name
        - email
        - message
      properties:
        name:
          type: string
        email:
          type: string
        subject:
          type: string
        message:
          type: string
        website:
          type: string
          description: Honeypot field — must be left empty by real users
    ContactResult:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
        message:
          type: string
    SubscribeResult:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
        message:
          type: string
    ErrorResult:
      type: object
      required:
        - error
      properties:
        error:
          type: string